backend/requirements.txt
# ============================================
# FastAPI Backend - Python 의존성
# ============================================

# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
motor==3.3.2                    # MongoDB 비동기 드라이버
redis==5.0.1                    # Redis 클라이언트

# Elasticsearch
elasticsearch[async]==8.11.0

# Kafka
aiokafka==0.10.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
httpx==0.26.0                   # HTTP 클라이언트
python-dotenv==1.0.0
ccxt==4.0.0                     # 암호화폐 거래소 API Client

#OCR
google-cloud-vision==3.7.4
google-auth==2.33.0
